backend/requirements.txt
django
djangorestframework
markdown
pillow
django-nested-inline